Conflicts:

	ChangeLog
	conf.c
	doc/Makefile.am
	doc/manual.html
	omfwd.c
	plugins/omgssapi/omgssapi.c

This was a bit hard to merge; if there are problems, they
may be in the area of the new "comment in action line" code
that came from the beta.

+++ runtime/nsd_gtls.c
@@ -500,7 +500,7 @@ gtlsRecordRecv(nsd_gtls_t *pThis)
 	}
 
 finalize_it:
-	dbgprintf("gtlsRecordRecv return. nsd %p, iRet %d, lenRcvd %d, lenRcvBuf %d, ptrRcvBuf %d\n", pThis, iRet, lenRcvd, pThis->lenRcvBuf, pThis->ptrRcvBuf);
+	dbgprintf("gtlsRecordRecv return. nsd %p, iRet %d, lenRcvd %d, lenRcvBuf %d, ptrRcvBuf %d\n", pThis, iRet, (int) lenRcvd, pThis->lenRcvBuf, pThis->ptrRcvBuf);
 	RETiRet;
 }
